pyproject.toml
[tool.poetry]
name = "ga-progression-modeling"
version = "0.1.0"
description = ""
authors = ["Aveturi13 <advaithveturi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.0.1"
pandas = "^2.2.2"
torch = "^2.4.0"
torchvision = "^0.19.0"
matplotlib = "^3.9.1"
umap-learn = "^0.5.6"
scikit-learn = "^1.5.1"
tqdm = "^4.66.4"
wandb = "^0.17.5"
jupyter = "^1.0.0"
opencv-python = "^4.10.0.84"
seaborn = "^0.13.2"
plotly = "^5.23.0"
dash = "^2.17.1"
scikit-image = "^0.24.0"
openpyxl = "^3.1.5"
kornia = "^0.7.3"
kornia-rs = "^0.1.5"
moviepy = "^1.0.3"
python-pptx = "^1.0.2"
monai = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
